+++ clevertap-core/src/main/java/com/clevertap/android/sdk/InboxMessageListener.java
@@ -2,7 +2,7 @@ package com.clevertap.android.sdk;
 
 import com.clevertap.android.sdk.inbox.CTInboxMessage;
 
-public interface CTInboxMessageListener {
+public interface InboxMessageListener {
     /**
      * callback to transfer payload when an inbox item is clicked
      */